Subscriptions: load widget from the Subcriptions module
Fixes https://github.com/Automattic/wp-calypso/issues/102467

This also has other advantages:

* It gets rid of the additional code in the muwpcom plugin.
* It keeps the main Jetpack class clean of unrelated code.
* It brings the dashboard widget PHP code close to its JS counterpart.
